+++ CHANGELOG.md
@@ -1,3 +1,17 @@
-## 0.0.1
+# Changelog
 
-* TODO: Describe initial release.
+## 0.0.1 - 2025-07-04
+
+- Initial release of the `app_guard` plugin.
+- Added methods for:
+  - `getPlatformVersion`
+  - `isDeviceSafe`
+  - `isDeviceRooted`
+  - `isDeveloperModeEnabled`
+  - `isDebuggingModeEnable`
+  - `isEmulator`
+  - `enableScreenshot`
+  - `disableScreenshot`
+  - `isDebuggerAttached`
+  - `isAppCloned`
+- Implemented basic Android and iOS native security checks.